+++ src/ralph/models/xapi/page_viewed.py
@@ -1,0 +1,19 @@
+"""PageViewed xAPI event definition"""
+
+from .base import BaseXapiModel
+from .fields.object import PageObjectField
+from .fields.verb import PageViewedVerbField
+
+
+class PageViewed(BaseXapiModel):
+    """Represents a page viewed xAPI statement.
+
+    Example: John viewed the https://www.fun-mooc.fr/ page.
+
+    Attributes:
+       object (PageObjectField): See PageObjectField.
+       verb (PageViewedVerbField): See PageViewedVerbField.
+    """
+
+    object: PageObjectField
+    verb: PageViewedVerbField = PageViewedVerbField()