+++ tests/test_utils.py
@@ -1,6 +1,8 @@
 """Tests for Ralph utils"""
 
 
+import copy
+
 import pytest
 
 from ralph import utils as ralph_utils
@@ -98,7 +100,7 @@ def test_utils_get_instance_from_class():
 
 @pytest.mark.parametrize("path,value", [(["foo", "bar"], "bar_value")])
 def test_utils_get_dict_value_from_path_should_return_given_value(path, value):
-    """The get_dict_value_from_path function should return the value when it's present."""
+    """Tests the get_dict_value_from_path function should return the value when it's present."""
 
     dictionary = {"foo": {"bar": "bar_value"}}
     assert ralph_utils.get_dict_value_from_path(dictionary, path) == value
@@ -116,7 +118,55 @@ def test_utils_get_dict_value_from_path_should_return_given_value(path, value):
 def test_utils_get_dict_value_from_path_should_return_none_when_value_does_not_exists(
     path,
 ):
-    """The get_dict_value_from_path function should return None if the value is not found."""
+    """Tests the get_dict_value_from_path function should return None if the value is not found."""
 
     dictionary = {"foo": {"bar": "bar_value"}}
     assert ralph_utils.get_dict_value_from_path(dictionary, path) is None
+
+
+def test_utils_set_dict_value_from_path_creating_new_fields():
+    """Tests when the fields are not present, set_dict_value_from_path should add them."""
+
+    dictionary = {}
+    ralph_utils.set_dict_value_from_path(dictionary, ["foo", "bar"], "baz")
+    assert dictionary == {"foo": {"bar": "baz"}}
+
+
+def test_utils_set_dict_value_from_path_updating_fields():
+    """Test when the fields are present, set_dict_value_from_path should update them."""
+
+    dictionary = {"foo": {"bar": "bar_value"}}
+    ralph_utils.set_dict_value_from_path(dictionary, ["foo", "bar"], "baz")
+    assert dictionary == {"foo": {"bar": "baz"}}
+
+
+@pytest.mark.parametrize(
+    "path",
+    [["foo", "bar", "baz"], ["not_foo", "not_bar"], ["not_foo", "bar"]],
+)
+def test_utils_remove_dict_key_from_path_should_not_change_the_dict_when_key_does_not_exists(
+    path,
+):
+    """Tests that the remove_dict_key_from_path function keeps the dictionary unchanged."""
+
+    dictionary = {"foo": {"bar": "bar_value"}, "baz": "baz_value"}
+    dictionary_copy = copy.deepcopy(dictionary)
+    ralph_utils.remove_dict_key_from_path(dictionary, path)
+    assert dictionary == dictionary_copy
+
+
+@pytest.mark.parametrize(
+    "path,expected",
+    [
+        (["foo"], {"baz": "baz_value"}),
+        (["foo", "bar"], {"foo": {}, "baz": "baz_value"}),
+    ],
+)
+def test_utils_remove_dict_key_from_path_should_remove_the_key_at_the_path(
+    path, expected
+):
+    """Tests that the remove_dict_key_from_path removes the field at the path."""
+
+    dictionary = {"foo": {"bar": "bar_value"}, "baz": "baz_value"}
+    ralph_utils.remove_dict_key_from_path(dictionary, path)
+    assert dictionary == expected